+++ crates/core/executor/src/minimal/precompiles/edwards/add.rs
@@ -3,9 +3,9 @@ use sp1_jit::JitContext;
 
 pub unsafe fn edwards_add<E: EdwardsParameters + EllipticCurve>(
     ctx: &mut JitContext,
-    arg1: u32,
-    arg2: u32,
-) -> Option<u32> {
+    arg1: u64,
+    arg2: u64,
+) -> Option<u64> {
     crate::minimal::precompiles::ec::ec_add::<E>(ctx, arg1, arg2);
 
     None
