+++ crates/core/executor/src/minimal/precompiles/weierstrass/decompress.rs
@@ -5,8 +5,8 @@ use sp1_jit::JitContext;
 #[allow(clippy::extra_unused_type_parameters)]
 pub(crate) fn weierstrass_decompress_syscall<E: EllipticCurve>(
     _ctx: &mut JitContext,
-    _slice_ptr: u32,
-    _sign_bit: u32,
-) -> Option<u32> {
+    _slice_ptr: u64,
+    _sign_bit: u64,
+) -> Option<u64> {
     panic!("This method should be deprecated.");
 }
